requirements.txt
pandas==1.5.2
Pillow==9.4.0
streamlit==1.19.0
